Fix the issue of person and book not being listed
